// view/next-project/src/generated/model/buyReportInformation.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * NUTFes FinanSu API
 * FinanSu APIドキュメント
 * OpenAPI spec version: 2.0.0
 */
import type { BuyReportInformationStatus } from './buyReportInformationStatus';

/**
 * マイページで表示する購入報告の情報
 */
export interface BuyReportInformation {
  id?: number;
  buyReportName?: string;
  amount?: number;
  reportDate?: string;
  status?: BuyReportInformationStatus;
}