// view/next-project/src/generated/model/festivalItemWithReport.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * NUTFes FinanSu API
 * FinanSu APIドキュメント
 * OpenAPI spec version: 2.0.0
 */
import type { Total } from './total';
import type { BuyReportInformation } from './buyReportInformation';

export interface FestivalItemWithReport {
  festivalItemTotal?: Total;
  festivalItemName?: string;
  buyReports?: BuyReportInformation[];
}